Huge protocol change and contract commit
This protocol has been entirely rewritten from previous commits in a way that allows other protocols to easily integrate using two instructions.
Each protocol may now initialize a court, which stores indexing details, then is allowed to create unique and customizable disputes given interactions
between users of the protocol.
The contract ensures that tokens and incentives follow an initial configuration from the protocol - the protocol is responsible for forcing users to sign an interact CPI in addition to some IX on their own contract.
While all customizations have not yet been implemented for Grizzlython, the planning is there to continue.